Creating all controllers routes for user and  events, Fixing some user and events issues, first version of back and front integration

+++ AgendaBack/Repositories/EventRepository.cs
@@ -17,38 +17,23 @@ using AgendaBack;
 
 namespace AgendaBack.Repositories
 {
-    internal class EventRepository
+    public class EventRepository
     {
 
-        private static string stringConexao = "Server=labsoft.pcs.usp.br; Initial Catalog = db_16; User id=usuario_16; pwd=51068523824";
-        //private readonly string stringConexao = "Data source=MP\\SQLEXPRESS; Initial Catalog=Catalog; integrated security=true;";
-
-      
-
-   
-
-        //**
-        public static string ReadInInterval(int id, DateTime start, DateTime end)
+        private static string stringConexao = System.Environment.GetEnvironmentVariable("stringConexao");
+        public static List<Event> ReadInInterval(int id, DateTime start, DateTime end)
         {
-
             List<Event> listEvents = new();
-
-
             using (SqlConnection con = new SqlConnection(stringConexao))
             {
                 string querySelect = "SELECT * FROM Eventos";
-
                 con.Open();
-
                 SqlDataReader rdr;
-
                 using (SqlCommand cmd = new SqlCommand(querySelect, con))
                 {
                     rdr = cmd.ExecuteReader();
-
                     while (rdr.Read())
                     {
-
                         Models.Event evento = new()
                         {
                             Id = Convert.ToInt32(rdr[0]),
@@ -58,19 +43,13 @@ namespace AgendaBack.Repositories
                             End = Convert.ToDateTime(rdr[4]),
                             addedUsersString = rdr[5].ToString(),
                             addedUsers = new List<int>()
-
                         };
-
                         string[] ids = evento.addedUsersString.Split(',');
                         foreach (string i in ids)
                         {
 
                             evento.addedUsers.Add(int.Parse(i));
                         }
-
-
-
-
                         if (evento.addedUsers.Contains(id) && DateTime.Compare(evento.Start, start) > 0 && DateTime.Compare(evento.Start, end) < 0)
                         {
                             listEvents.Add(evento);
@@ -80,8 +59,7 @@ namespace AgendaBack.Repositories
                     }
                 }
             }
-
-            return JsonSerializer.Serialize(listEvents);
+            return listEvents;
         }
 
         
@@ -93,61 +71,50 @@ namespace AgendaBack.Repositories
        
 
         // **
-        public static void CreateEvent(int creatorId, int IdEvent, string Name, string Description, DateTime Start, DateTime End)
+        public static void CreateEvent(int creatorId, string Name, string Description, DateTime Start, DateTime End)
         {
 
             
             using (SqlConnection con = new SqlConnection(stringConexao))
             {
-
-
-                string queryInsert = "INSERT INTO Eventos (Id, [Name], [Description], [Start], [End], Users) VALUES (@Id, @Name, @Description, @Start, @End, @Users)";
-
+                string queryInsert = "INSERT INTO Eventos ([Name], [Description], [Start], [End], Users, OwnerId) VALUES (@Name, @Description, @Start, @End, @Users, @OwnerId)";
                 using (SqlCommand cmd = new SqlCommand(queryInsert, con))
                 {
-                    cmd.Parameters.AddWithValue("@Id", IdEvent);
                     cmd.Parameters.AddWithValue("@Name", Name);
                     cmd.Parameters.AddWithValue("@Description", Description);
                     cmd.Parameters.AddWithValue("@Start", Start);
                     cmd.Parameters.AddWithValue("@End",  End);
                     cmd.Parameters.AddWithValue("@Users", creatorId.ToString());
-
+                    cmd.Parameters.AddWithValue("@OwnerId", creatorId);
                     con.Open();
-
                     cmd.ExecuteNonQuery();
                 }
             }
         }
 
         //**
-        public static void DeleteEvent(int IdEvent)
+        public static void DeleteEvent(int IdEvent, int OwnerId)
         {
-
             using (SqlConnection con = new SqlConnection(stringConexao))
             {
-
-
-                string queryInsert = "DELETE FROM Eventos WHERE Id = @Id";
+                string queryInsert = "DELETE FROM Eventos WHERE Id = @Id and OwnerId = @OwnerId";
 
                 using (SqlCommand cmd = new SqlCommand(queryInsert, con))
                 {
                     cmd.Parameters.AddWithValue("@Id", IdEvent);
-                    
+                    cmd.Parameters.AddWithValue("@OwnerId", OwnerId);
                     con.Open();
-
                     cmd.ExecuteNonQuery();
                 }
             }
         }
 
         //**
-        public static void EditEvent(int IdEvent, string Name, string Description, DateTime Start, DateTime End)
+        public static void EditEvent(int IdEvent, string Name, string Description, DateTime Start, DateTime End, int OwnerId)
         {
             using (SqlConnection con = new SqlConnection(stringConexao))
             {
-
-
-                string queryInsert = "UPDATE Eventos SET [Name] = @Name, [Description] = @Description, [Start] = @Start, [End] = @End WHERE Id = @Id";
+                string queryInsert = "UPDATE Eventos SET [Name] = @Name, [Description] = @Description, [Start] = @Start, [End] = @End WHERE Id = @Id and OwnerId = @OwnerId";
 
                 using (SqlCommand cmd = new SqlCommand(queryInsert, con))
                 {
@@ -156,21 +123,12 @@ namespace AgendaBack.Repositories
                     cmd.Parameters.AddWithValue("@Description", Description);
                     cmd.Parameters.AddWithValue("@Start", Start);
                     cmd.Parameters.AddWithValue("@End", End);
-
-
+                    cmd.Parameters.AddWithValue("@OwnerId", OwnerId);
                     con.Open();
-
                     cmd.ExecuteNonQuery();
                 }
             }
-
         }
-
-
-
-
-
-
 
     }
 }
